+++ lib/fade_direction.dart
@@ -1,0 +1,17 @@
+/// An enum that defines the direction of the fade effect for the particles.
+enum FadeDirection {
+  /// Fade from the top of the canvas.
+  top,
+
+  /// Fade from the bottom of the canvas.
+  bottom,
+
+  /// Fade from the left of the canvas.
+  left,
+
+  /// Fade from the right of the canvas.
+  right,
+
+  /// No fade effect.
+  none,
+}